Move hmdStatusChange back to VRView; Fix Viewer deserialization logic for egde cases
