+++ Scripts/Core/EditorVR.cs
@@ -11,7 +11,6 @@ using UnityEditor.Experimental.EditorVR.Utilities;
 using UnityEngine;
 using UnityEngine.Assertions;
 using UnityEngine.InputNew;
-using UnityEngine.VR;
 
 namespace UnityEditor.Experimental.EditorVR.Core
 {
@@ -39,9 +38,6 @@ namespace UnityEditor.Experimental.EditorVR.Core
 
 		event Action m_SelectionChanged;
 
-		public static event Action<bool> hmdStatusChange;
-		bool m_HMDReady;
-
 		readonly List<DeviceData> m_DeviceData = new List<DeviceData>();
 
 		// Local method use only -- caching here to prevent frequent lookups in Update
@@ -54,6 +50,8 @@ namespace UnityEditor.Experimental.EditorVR.Core
 		DeviceInputModule m_DeviceInputModule;
 		Viewer m_Viewer;
 		MultipleRayInputModule m_MultipleRayInputModule;
+
+		bool m_HasDeserialized;
 
 		static HideFlags defaultHideFlags
 		{
@@ -133,7 +131,6 @@ namespace UnityEditor.Experimental.EditorVR.Core
 
 			m_Viewer = GetNestedModule<Viewer>();
 			m_Viewer.preserveCameraRig = preserveLayout;
-			hmdStatusChange += m_Viewer.OnHMDStatusChange;
 			m_Viewer.InitializeCamera();
 
 			var tools = GetNestedModule<Tools>();
@@ -242,10 +239,12 @@ namespace UnityEditor.Experimental.EditorVR.Core
 				yield return null;
 			}
 
-			while (!m_HMDReady)
+			var viewer = GetNestedModule<Viewer>();
+			while (!viewer.hmdReady)
 				yield return null;
 
 			GetModule<SerializedPreferencesModule>().DeserializePreferences(serializedPreferences);
+			m_HasDeserialized = true;
 		}
 
 		static void ClearDeveloperConsoleIfNecessary()
@@ -302,13 +301,12 @@ namespace UnityEditor.Experimental.EditorVR.Core
 
 		void Shutdown()
 		{
-			serializedPreferences = GetModule<SerializedPreferencesModule>().SerializePreferences();
+			if (m_HasDeserialized)
+				serializedPreferences = GetModule<SerializedPreferencesModule>().SerializePreferences();
 		}
 
 		void OnDestroy()
 		{
-			hmdStatusChange -= m_Viewer.OnHMDStatusChange;
-
 			foreach (var nested in m_NestedModules.Values)
 			{
 				nested.OnDestroy();
@@ -317,7 +315,6 @@ namespace UnityEditor.Experimental.EditorVR.Core
 
 		void Update()
 		{
-			UpdateHMDStatus();
 			m_Viewer.UpdateCamera();
 
 			m_Rays.UpdateRaycasts();
@@ -332,19 +329,6 @@ namespace UnityEditor.Experimental.EditorVR.Core
 			m_Menus.UpdateMenuVisibilities();
 
 			m_UI.UpdateManipulatorVisibilites();
-		}
-
-		void UpdateHMDStatus()
-		{
-			if (hmdStatusChange != null)
-			{
-				var ready = GetIsUserPresent();
-				if (m_HMDReady != ready)
-				{
-					m_HMDReady = ready;
-					hmdStatusChange(ready);
-				}
-			}
 		}
 
 		void ProcessInput(HashSet<IProcessInput> processedInputs, ConsumeControlDelegate consumeControl)
@@ -519,19 +503,6 @@ namespace UnityEditor.Experimental.EditorVR.Core
 			ObjectUtils.hideFlags = defaultHideFlags;
 			InitializeInputManager();
 			s_Instance = ObjectUtils.CreateGameObjectWithComponent<EditorVR>();
-		}
-
-		static bool GetIsUserPresent()
-		{
-#if ENABLE_OVR_INPUT
-			if (VRSettings.loadedDeviceName == "Oculus")
-				return OVRPlugin.userPresent;
-#endif
-#if ENABLE_STEAMVR_INPUT
-			if (VRSettings.loadedDeviceName == "OpenVR")
-				return OpenVR.System.GetTrackedDeviceActivityLevel(0) == EDeviceActivityLevel.k_EDeviceActivityLevel_UserInteraction;
-#endif
-			return true;
 		}
 
 		static void InitializeInputManager()
